feature28-jwt: move files into correct package folder
