+++ src/main/java/com/example/myinsta/controller/JwtController.java
@@ -1,5 +1,9 @@
-package com.example.myinsta.security;
+package com.example.myinsta.controller;
 
+import com.example.myinsta.service.JwtService;
+import com.example.myinsta.dto.LoginDto;
+import com.example.myinsta.dto.RequestSignUpDto;
+import com.example.myinsta.dto.ResponseSignInDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,8 +25,8 @@ public class JwtController {
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
     @PostMapping("/jwt/signIn")
-    public ResponseEntity<SignInResponseDto> signIn(@Valid @RequestBody LoginDto loginDTO) {
-        SignInResponseDto signInResponseDto = jwtService.login(loginDTO);
-        return ResponseEntity.status(OK).body(signInResponseDto);
+    public ResponseEntity<ResponseSignInDto> signIn(@Valid @RequestBody LoginDto loginDTO) {
+        ResponseSignInDto responseSignInDto = jwtService.login(loginDTO);
+        return ResponseEntity.status(OK).body(responseSignInDto);
     }
 }
