+++ src/main/java/com/example/myinsta/dto/ResponseSignInDto.java
@@ -1,4 +1,4 @@
-package com.example.myinsta.security;
+package com.example.myinsta.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 @Data
 @AllArgsConstructor
-public class SignInResponseDto {
+public class ResponseSignInDto {
     private String grantType;
     private String accessToken;
 }
