+++ src/main/java/com/example/myinsta/mapper/JwtMapper.java
@@ -1,6 +1,6 @@
 package com.example.myinsta.mapper;
 
-import com.example.myinsta.security.AccountsDao;
+import com.example.myinsta.dao.AccountsDao;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
@@ -10,5 +10,5 @@ public interface JwtMapper {
     void addRole(AccountsDao.RoleDao roleDao);
     AccountsDao.AccountDetailsDao findByEmailPassword(String email, String password);
     AccountsDao.AccountDetailsDao findByEmail(String email);
-
+    boolean isIdExist(AccountsDao.AccountDetailsDao accountDetailsDao);
 }
