+++ src/main/java/com/example/myinsta/security/JwtExceptionHandler.java
@@ -1,6 +1,7 @@
 package com.example.myinsta.security;
 
 import com.example.myinsta.exception.ErrorResponse;
+import com.example.myinsta.exception.JwtCustomException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
