+++ src/main/java/com/example/myinsta/security/JwtTokenFilter.java
@@ -1,5 +1,6 @@
 package com.example.myinsta.security;
 
+import com.example.myinsta.exception.JwtCustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
