+++ src/main/java/com/example/myinsta/service/CustomDetailService.java
@@ -1,5 +1,6 @@
-package com.example.myinsta.security;
+package com.example.myinsta.service;
 
+import com.example.myinsta.dao.AccountsDao;
 import com.example.myinsta.mapper.JwtMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.User;
