+++ src/main/resources/mapper/JwtMapper.xml
@@ -1,26 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.example.myinsta.mapper.JwtMapper">
-    <insert id="signUpAccount" parameterType="com.example.myinsta.security.AccountsDao$AccountDetailsDao" useGeneratedKeys="true" keyProperty="idAccount">
+    <insert id="signUpAccount" parameterType="com.example.myinsta.dao.AccountsDao$AccountDetailsDao" useGeneratedKeys="true" keyProperty="idAccount">
         insert into accounts( email, nick_name, password )
         values( #{email}, #{nickName}, #{password} )
     </insert>
-    <insert id="signUpRole" parameterType="com.example.myinsta.security.AccountsDao$RoleDao">
+    <insert id="signUpRole" parameterType="com.example.myinsta.dao.AccountsDao$RoleDao">
         insert into authority ( id_account, role )
         values ( #{idAccount}, #{role} )
     </insert>
-    <insert id = "addRole" parameterType="com.example.myinsta.security.AccountsDao$RoleDao">
+    <insert id = "addRole" parameterType="com.example.myinsta.dao.AccountsDao$RoleDao">
         insert into authority ( id_account, role )
         values ( #{idAccount}, #{role} )
     </insert>
-    <resultMap id="memberInfo" type="com.example.myinsta.security.AccountsDao$AccountDetailsDao">
+    <resultMap id="memberInfo" type="com.example.myinsta.dao.AccountsDao$AccountDetailsDao">
         <id column="id_account" property="idAccount"/>
         <result column="email" property="email"/>
         <result column="nick_name" property="nickName"/>
         <result column="password" property="password"/>
         <association property="role" resultMap="roleInfo"/>
     </resultMap>
-    <resultMap id="roleInfo" type="com.example.myinsta.security.AccountsDao$RoleDao">
+    <resultMap id="roleInfo" type="com.example.myinsta.dao.AccountsDao$RoleDao">
         <id column="id_authority" property="idAuthority"/>
         <result column="role" property="role"/>
     </resultMap>
@@ -36,5 +36,11 @@
         select A.id_account, A.email, A.password, A.nick_name, B.role
         from accounts as A LEFT JOIN authority as B ON A.id_account = B.id_account
         where email = #{email}
+    </select>
+    <select id="isIdExist" parameterType="com.example.myinsta.dao.AccountsDao$AccountDetailsDao" resultType="boolean">
+        SELECT EXISTS
+        (SELECT email
+        FROM accounts
+        WHERE email = #{email});
     </select>
 </mapper>
