+++ src/test/java/com/example/myinsta/controller/AccountsControllerTest.java
@@ -1,7 +1,7 @@
 package com.example.myinsta.controller;
 
 import com.example.myinsta.dto.RequestSignUpDto;
-import com.example.myinsta.service.AccountsService;
+import com.example.myinsta.service.JwtService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -48,7 +48,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  */
 
 
-@WebMvcTest(controllers = AccountsController.class)
+@WebMvcTest(controllers = JwtController.class)
 public class AccountsControllerTest {
 
     @Autowired
@@ -62,7 +62,7 @@ public class AccountsControllerTest {
     ObjectMapper objectMapper;
 
     @MockBean
-    AccountsService accountsService;
+    JwtService accountsService;
 
     @BeforeEach
     void setUp() {
